Move Estimator to estimators package  * Move SubGraph to libs  * Move Summarizer to models
