+++ polyaxon/experiments/__init__.py
@@ -1,7 +1,4 @@
 # -*- coding: utf-8 -*-
 from __future__ import absolute_import, division, print_function
 
-from polyaxon.experiments.hooks import HOOKS
-from polyaxon.experiments.estimator import Estimator, ESTIMATORS
 from polyaxon.experiments.experiment import Experiment, create_experiment, run_experiment
-from polyaxon.experiments.subgraph import SubGraph
