+++ polyaxon/libs/__init__.py
@@ -5,3 +5,4 @@ from .collections import *
 from . import getters
 from . import configs
 from . import utils
+from .subgraph import SubGraph
